+++ ExampleProject/platforms/android/app/src/main/assets/www/cordova_plugins.js
@@ -11,6 +11,6 @@ cordova.define('cordova/plugin_list', function(require, exports, module) {
   ];
   module.exports.metadata = {
     "cordova-plugin-whitelist": "1.3.4",
-    "clevertap-cordova": "2.3.0"
+    "clevertap-cordova": "2.3.2"
   };
 });