+++ ExampleProject/platforms/android/app/src/main/assets/www/js/index.js
@@ -37,8 +37,6 @@ initialize: function() {
     document.addEventListener('onCleverTapProductConfigDidActivate', this.onCleverTapProductConfigDidActivate, false);
     document.addEventListener('onCleverTapExperimentsUpdated', this.onCleverTapExperimentsUpdated, false);
     document.addEventListener('onCleverTapDisplayUnitsLoaded', this.onCleverTapDisplayUnitsLoaded, false);
-    document.addEventListener('onCleverTapNotificationClickedPayloadDidReceived', this.onCleverTapNotificationClickedPayloadDidReceived, false);
-    document.addEventListener('onCleverTapPushAmpPayloadDidReceived', this.onCleverTapPushAmpPayloadDidReceived, false);
 },
     
     // deviceready Event Handler
@@ -50,6 +48,8 @@ onDeviceReady: function() {
     CleverTap.setDebugLevel(3);
     CleverTap.registerPush();
     
+    CleverTap.recordEventWithNameAndProps("foo", {"bar":"zoo"});
+
     // Ionic example usage
     //$rootScope.CleverTap = CleverTap;
     //CleverTap && CleverTap.registerPush();
@@ -194,6 +194,10 @@ onDeviceReady: function() {
     //      }
     //
     //
+
+onCleverTapPushNotificationTappedWithCustomExtras: function(e){
+    console.log(e.customExtras);
+}
     
 onCleverTapProfileSync: function(e) {
     console.log(e.updates);
@@ -260,16 +264,6 @@ onCleverTapExperimentsUpdated: function() {
 onCleverTapDisplayUnitsLoaded: function(e) {
     console.log("onCleverTapDisplayUnitsLoaded");
     console.log(e.units);
-},
-
-onCleverTapNotificationClickedPayloadDidReceived: function(e) {
-    console.log("onCleverTapNotificationClickedPayloadDidReceived");
-    console.log(e.customExtras);
-},
-
-onCleverTapPushAmpPayloadDidReceived: function(e) {
-    console.log("onCleverTapPushAmpPayloadDidReceived");
-    console.log(e.customExtras);
 },
     
     // Update DOM on a Received Event