+++ ExampleProject/platforms/ios/Pods/Headers/Private/CleverTap-iOS-SDK/CTABTestEditorSnapshotMessageResponse.h
@@ -1,0 +1,1 @@
+../../../CleverTap-iOS-SDK/CleverTapSDK/ABTesting/uieditor/messages/CTABTestEditorSnapshotMessageResponse.h