+++ ExampleProject/platforms/ios/Pods/Headers/Private/CleverTap-iOS-SDK/CTFeatureFlagsController.h
@@ -1,0 +1,1 @@
+../../../CleverTap-iOS-SDK/CleverTapSDK/FeatureFlags/controllers/CTFeatureFlagsController.h