+++ ExampleProject/platforms/ios/www/js/index.js
@@ -37,6 +37,8 @@ initialize: function() {
     document.addEventListener('onCleverTapProductConfigDidActivate', this.onCleverTapProductConfigDidActivate, false);
     document.addEventListener('onCleverTapExperimentsUpdated', this.onCleverTapExperimentsUpdated, false);
     document.addEventListener('onCleverTapDisplayUnitsLoaded', this.onCleverTapDisplayUnitsLoaded, false);
+    document.addEventListener('onCleverTapPushNotificationTappedWithCustomExtras', this.onCleverTapPushNotificationTappedWithCustomExtras, false);
+    document.addEventListener('onCleverTapPushAmpPayloadDidReceived', this.onCleverTapPushAmpPayloadDidReceived, false);
 },
     
     // deviceready Event Handler
@@ -45,8 +47,6 @@ initialize: function() {
     // 'pause', 'resume', etc.
 onDeviceReady: function() {
     this.receivedEvent('deviceready');
-    CleverTap.setDebugLevel(3);
-    CleverTap.registerPush();
     
     // Ionic example usage
     //$rootScope.CleverTap = CleverTap;
@@ -205,7 +205,12 @@ onCleverTapInAppNotificationDismissed: function(e) {
     console.log(e.extras);
     console.log(e.actionExtras);
 },
-    
+
+onCleverTapPushNotificationTappedWithCustomExtras: function(e) {
+    console.log("onCleverTapPushNotificationTappedWithCustomExtras");
+    console.log(e.customExtras);
+}, 
+
     // deep link handling
 onDeepLink: function(e) {
     console.log(e.deeplink);
@@ -259,7 +264,16 @@ onCleverTapDisplayUnitsLoaded: function(e) {
     console.log("onCleverTapDisplayUnitsLoaded");
     console.log(e.units);
 },
-    
+onCleverTapPushNotificationTappedWithCustomExtras: function(e) {
+    console.log("onCleverTapPushNotificationTappedWithCustomExtras");
+    console.log(e.customExtras);
+},
+
+onCleverTapPushAmpPayloadDidReceived: function(e) {
+    console.log("onCleverTapPushAmpPayloadDidReceived");
+    console.log(e.customExtras);
+},
+
     // Update DOM on a Received Event
 receivedEvent: function(id) {
     var parentElement = document.getElementById(id);