+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 Change Log
 ==========
 
-Version 2.3.2 *(11 December, 2020)*
+Version 2.3.2 *(06 January, 2021)*
 -------------------------------------------
-* Added support for Notification Click Callbacks.
+* Added support for Push Notification Click Callbacks.
+* Added CleverTap iOS SDK dependency via CocoaPods in the plugin.xml
 * Update for CleverTap Android SDK v4.0.1
 
 Version 2.3.1 *(11 November, 2020)*